feat: generate json schema
This commit introduces the ability to generate a json schema for
concourse pipeline definitions.

The schema is generated from the data structures defined in the
src/schema directory.

The schema is serialized to a string and written to stdout.

The schema is based on draft-04 of the json schema specification.